# Conflicts:
#	app/src/main/java/com/keylesspalace/tusky/di/FragmentBuildersModule.kt
#	app/src/main/java/com/keylesspalace/tusky/util/ThrowableExtensions.kt

+++ app/src/main/java/com/keylesspalace/tusky/components/filters/FiltersActivity.kt
@@ -1,5 +1,6 @@
 package com.keylesspalace.tusky.components.filters
 
+import android.content.DialogInterface.BUTTON_POSITIVE
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
@@ -60,13 +61,13 @@ class FiltersActivity : BaseActivity(), FiltersListener {
                 when (state.loadingState) {
                     FiltersViewModel.LoadingState.INITIAL, FiltersViewModel.LoadingState.LOADING -> binding.messageView.hide()
                     FiltersViewModel.LoadingState.ERROR_NETWORK -> {
-                        binding.messageView.setup(R.drawable.elephant_offline, R.string.error_network) {
+                        binding.messageView.setup(R.drawable.errorphant_offline, R.string.error_network) {
                             loadFilters()
                         }
                         binding.messageView.show()
                     }
                     FiltersViewModel.LoadingState.ERROR_OTHER -> {
-                        binding.messageView.setup(R.drawable.elephant_error, R.string.error_generic) {
+                        binding.messageView.setup(R.drawable.errorphant_error, R.string.error_generic) {
                             loadFilters()
                         }
                         binding.messageView.show()
@@ -104,7 +105,11 @@ class FiltersActivity : BaseActivity(), FiltersListener {
     }
 
     override fun deleteFilter(filter: Filter) {
-        viewModel.deleteFilter(filter, binding.root)
+        lifecycleScope.launch {
+            if (showDeleteFilterDialog(filter.title) == BUTTON_POSITIVE) {
+                viewModel.deleteFilter(filter, binding.root)
+            }
+        }
     }
 
     override fun updateFilter(updatedFilter: Filter) {
